Progress Save - Async Delete/SQS
Have to go run errands/do chores, saving for progress.

Small refactors, moving to a soft-delete until ASYNC hard delete
pattern as AWS IoT requires several steps prior to a delete to be done
in order, with one step being completely async with up to 3 seconds lag.

Pending-delete attribute is added to items and queries updated to not
display anything with PENDING_DELETE.